requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
crispy-bootstrap5==0.6
Django==4.0.4
django-crispy-forms==1.14.0
Pillow==9.1.0
sqlparse==0.4.2
